+++ CHANGELOG.md
@@ -22,6 +22,8 @@ if they are different, the hash will be highlighted in red
 * Repeated T will close the tags edit flyout
 * Escape to exit from flyouts
 * Autofocus search tag field in tags edit
+* Fix the bug in quick add tags when different sets have the same key
+* Add ctrl shift space hotkey for quick tagging to select the previous tag pack set
 
 ### Viewer
 * Support opening zipped galleries: zip 7z rar cbz cbr etc It will be extracted to a temporary folder and opened